refactor(meta): move `pause`/`resume` to stream service
