+++ src/meta/src/rpc/service/scale_service.rs
@@ -16,22 +16,20 @@ use risingwave_pb::common::WorkerType;
 use risingwave_pb::meta::scale_service_server::ScaleService;
 use risingwave_pb::meta::{
     GetClusterInfoRequest, GetClusterInfoResponse, GetReschedulePlanRequest,
-    GetReschedulePlanResponse, PauseRequest, PauseResponse, Reschedule, RescheduleRequest,
-    RescheduleResponse, ResumeRequest, ResumeResponse,
+    GetReschedulePlanResponse, Reschedule, RescheduleRequest, RescheduleResponse,
 };
 use risingwave_pb::source::{ConnectorSplit, ConnectorSplits};
 use tonic::{Request, Response, Status};
 
-use crate::barrier::{BarrierManagerRef, BarrierScheduler, Command};
+use crate::barrier::BarrierManagerRef;
 use crate::manager::{CatalogManagerRef, ClusterManagerRef, FragmentManagerRef};
-use crate::model::{MetadataModel, PausedReason};
+use crate::model::MetadataModel;
 use crate::storage::MetaStore;
 use crate::stream::{
     GlobalStreamManagerRef, ParallelUnitReschedule, RescheduleOptions, SourceManagerRef,
 };
 
 pub struct ScaleServiceImpl<S: MetaStore> {
-    barrier_scheduler: BarrierScheduler<S>,
     fragment_manager: FragmentManagerRef<S>,
     cluster_manager: ClusterManagerRef<S>,
     source_manager: SourceManagerRef<S>,
@@ -45,7 +43,6 @@ where
     S: MetaStore,
 {
     pub fn new(
-        barrier_scheduler: BarrierScheduler<S>,
         fragment_manager: FragmentManagerRef<S>,
         cluster_manager: ClusterManagerRef<S>,
         source_manager: SourceManagerRef<S>,
@@ -54,7 +51,6 @@ where
         barrier_manager: BarrierManagerRef<S>,
     ) -> Self {
         Self {
-            barrier_scheduler,
             fragment_manager,
             cluster_manager,
             source_manager,
@@ -70,26 +66,6 @@ impl<S> ScaleService for ScaleServiceImpl<S>
 where
     S: MetaStore,
 {
-    #[cfg_attr(coverage, no_coverage)]
-    async fn pause(&self, _: Request<PauseRequest>) -> Result<Response<PauseResponse>, Status> {
-        // TODO: move this out of the scale service, as scaling actually executes `pause` and
-        // `resume` with `PausedReason::ConfigChange`.
-        self.barrier_scheduler
-            .run_command(Command::pause(PausedReason::Manual))
-            .await?;
-        Ok(Response::new(PauseResponse {}))
-    }
-
-    #[cfg_attr(coverage, no_coverage)]
-    async fn resume(&self, _: Request<ResumeRequest>) -> Result<Response<ResumeResponse>, Status> {
-        // TODO: move this out of the scale service, as scaling actually executes `pause` and
-        // `resume` with `PausedReason::ConfigChange`.
-        self.barrier_scheduler
-            .run_command(Command::resume(PausedReason::Manual))
-            .await?;
-        Ok(Response::new(ResumeResponse {}))
-    }
-
     #[cfg_attr(coverage, no_coverage)]
     async fn get_cluster_info(
         &self,
